backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
psycopg2-binary
anthropic
